+++ NEWS.md
@@ -14,7 +14,9 @@
     - `Bool` in `markup` mode is no longer prefixed by `#`
     - An `AbstractChar` and `AbstractString` in `code` and `math` mode are no longer escaped and quoted
     - A `Complex` is no longer parenthesized when a term is zero
-- The `show` methods for `TypstString` with a `application/pdf`, `image/png`, and `image/svg+xml` MIME types now support a custom `preamble`
+- The `show` methods with the `applicatoin/pdf`, `image/png`, and `image/svg+xml` MIME types
+    - Can be given a custom `preamble`
+    - Support `Typst` and `TypstText`
 
 ### Commands